exercícios de introdução ao java

+++ IntroducaoAoJava/src/introducaoaojava/desvioscondicionais/Exercicio02.java
@@ -6,7 +6,7 @@ import java.util.Scanner;
 public class Exercicio02 {public static void main(String[] args) {
         Scanner entrada = new Scanner(System.in);
             System.out.println("Informe um numero qualquer");
-            int numero =entrada.nextInt();
+            double numero =entrada.nextDouble();
             
            if(numero %2==0){
             System.out.println("É par");
